+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # CHANGELOG
+* `3.6.0 `Release - [3.6.0](#3.6.0)&nbsp; - &nbsp; 2023-04-13
 * `3.5.5 `Release - [3.5.5](#3.5.5)&nbsp; - &nbsp; 2023-03-30
 * `3.5.4 `Release - [3.5.4](#3.5.4)&nbsp; - &nbsp; 2023-03-17
 * `3.5.3 `Release - [3.5.3](#3.5.3)&nbsp; - &nbsp; 2023-03-15
@@ -35,6 +36,15 @@
 * `3.1.8` Release - [3.1.8](#3.1.8) &nbsp; - &nbsp; 2020-12-11
 * `3.1.7` Release - [3.1.7](#3.1.7) &nbsp; - &nbsp; 2020-12-04
 * `3.1.6` Release - [3.1.6](#3.1.6) &nbsp; - &nbsp; 2020-11-13
+
+# <a name="3.6.0"></a> 3.6.0
+#### Features
+`
+DevEnigmaRiverAndroidCore
+DevEnigmaRiverAndroidInternalTestUtils
+`
+* EMP-19633 - Provide device info as ad-params in play-requests
+* EMP-19409 - Iterate over SSAI parameters sent in the play request
 
 # <a name="3.5.5"></a> 3.5.5
 #### Features